change int to long

+++ WMS_VS2015_NO/AplombTech.WMS/Domain/AplombTech.WMS.Domain/SummaryData/OnOffData.cs
@@ -13,7 +13,7 @@ namespace AplombTech.WMS.Domain.SummaryData
         #region Primitive Properties
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public virtual int OnOffDataId { get; set; }
+        public virtual Int64 OnOffDataId { get; set; }
         public virtual DateTime OffDateTime { get; set; }
         public virtual DateTime? OnDateTime { get; set; }
         public virtual double Duration { get; set; }
