+++ WMS_VS2015_NO/AplombTech.WMS/Domain/AplombTech.WMS.QueryModel/SummaryData/OnOffData.cs
@@ -12,7 +12,7 @@ namespace AplombTech.WMS.QueryModel.SummaryData
     {
         #region Primitive Properties
         [Key]
-        public virtual int OnOffDataId { get; set; }
+        public virtual Int64 OnOffDataId { get; set; }
         public virtual DateTime OffDateTime { get; set; }
         public virtual DateTime? OnDateTime { get; set; }
         public virtual double Duration { get; set; }
